ConstantKernel bug + minor typos
Multiplying by a constant normally adds an extra parameter from the ConstantKernel, which breaks the code; we need to call ConstantKernel directly and specify that this should be an immutable parameter.

Also removed the frivolous differences between rmjarvis/treegp (which this was originally cloned from) and pfleget/treegp (which this is being pushed to.